[GREEN] Just pass the test!
We don't need to do much to just pass the test. We can just return
static content. It is really easy to write more code than is actually
needed, but it holds the risk of not testing everything.

If you want to write more code, first write a new test.