[RED/GREEN] Calculate homework average for a single homework
Again we need to resist writing too much code. We know that there will
be multiple homework scores, but for now, we can calculate the average
only for single homework.

[BACKWARD COMPATIBILTY] We wrote much more code because we wanted to be
backwards compatible with an old interface. That's overkill for such a
simple example, but in a complex codebase, it could save us from
modifying tons of files (and merge hell).